decompile: upstream 8fdeda2b9b6e94761ab7fead714391e6bd27d486 (2021-09-09) -- 230 files changed
Upstream commit: https://github.com/espressif/esp-thread-lib/commit/8fdeda2b9b6e94761ab7fead714391e6bd27d486

Upstream subject: br: fix router solicitation handling(e82fe0d)

Archives decompiled: 2

+++ decompiled/libopenthread_port/esp_openthread_task_queue/esp_openthread_task_queue_update.c
@@ -1,8 +1,8 @@
 /*
- * Last changed at upstream commit 852eceaf38f6d6304f3b446a4a26f861389f83be
- * https://github.com/espressif/esp-thread-lib/commit/852eceaf38f6d6304f3b446a4a26f861389f83be
- * Upstream date: 2021-07-06 14:49:24 +0800
- * Upstream subject: openthread: make queue size and partition configurable
+ * Last changed at upstream commit 8fdeda2b9b6e94761ab7fead714391e6bd27d486
+ * https://github.com/espressif/esp-thread-lib/commit/8fdeda2b9b6e94761ab7fead714391e6bd27d486
+ * Upstream date: 2021-09-09 20:40:32 +0800
+ * Upstream subject: br: fix router solicitation handling(e82fe0d)
  * Source: libopenthread_port -> esp_openthread_task_queue.o -> esp_openthread_task_queue_update
  *
  * (C) Espressif, Apache License 2.0.
@@ -18,7 +18,7 @@ void esp_openthread_task_queue_update(int param_1)
   
   uVar1 = s_task_queue_event_fd;
   if (-1 < (int)s_task_queue_event_fd) {
-    if (s_task_queue_event_fd < 0x40) {
+    if ((int)s_task_queue_event_fd < 0x40) {
       puVar2 = (uint *)((s_task_queue_event_fd >> 5) * 4 + param_1);
       *puVar2 = *puVar2 | 1 << (s_task_queue_event_fd & 0x1f);
     }
